Live Project
This is all the code that was used in creating a project during a Live Project using Azure DevOps in a team setting. It is an application that allows users to enter items into a database, as well as allows them to view those items, edit them, and even delete them.